[Fix] 선택 데이터 기존 데이터에서 필터
