+++ src/features/accountItem/index.ts
@@ -1,3 +1,2 @@
 export type { AccountItem, MoneyType } from './model/types'
-export { getByMonth, getByDate } from './model/queries'
-export { fetchByMonth, fetchByDate } from './service/accountItem'
+export { fetchByMonth } from './service/accountItem'
