--- src/features/accountItem/model/queries.ts
@@ -1,4 +1,0 @@
-import { fetchByMonth, fetchByDate } from '../service/accountItem'
-
-export const getByMonth = fetchByMonth
-export const getByDate = fetchByDate
