+++ src/features/calendar/index.ts
@@ -1,5 +1,5 @@
 export { Calendar } from './ui/calendar/Calendar'
 export { PickDate } from './ui/PickDate'
 export { DateListOverlay } from './ui/overlay/DateListOverlay'
-export { useCalendar } from './model/useCalendar'
+
 export { useSelectedDate } from './model/useSelectedDate'
